combined the 2 alphabetical controllers in 1 to remove duplication
