+++ src/main/routes/routes.ts
@@ -115,8 +115,8 @@ export default function(app: Application): void {
   app.get('/case-reference-number-search', ensureAuthenticated, app.locals.container.cradle.caseReferenceNumberSearchController.get);
   app.post('/case-reference-number-search', ensureAuthenticated, app.locals.container.cradle.caseReferenceNumberSearchController.post);
   app.get('/case-reference-number-search-results', ensureAuthenticated, app.locals.container.cradle.caseReferenceNumberSearchResultController.get);
-  app.get('/court-name-search', ensureAuthenticated, app.locals.container.cradle.courtNameSearchController.get);
-  app.post('/court-name-search', ensureAuthenticated, app.locals.container.cradle.courtNameSearchController.post);
+  app.get('/court-name-search', ensureAuthenticated, app.locals.container.cradle.alphabeticalSearchController.get);
+  app.post('/court-name-search', ensureAuthenticated, app.locals.container.cradle.alphabeticalSearchController.post);
   app.get('/create-admin-account', ensureAuthenticated, app.locals.container.cradle.createAdminAccountController.get);
   app.post('/create-admin-account', ensureAuthenticated, app.locals.container.cradle.createAdminAccountController.post);
   app.get('/create-admin-account-summary', ensureAuthenticated, app.locals.container.cradle.createAdminAccountSummaryController.get);
